파일 경로 수정 및 시나리오 엔티티, 레포지토리 추가
