--- backend/scenario/src/main/java/com/sk/skala/axcalibur/scenario/apilist/global/entity/FilePathEntity.java
@@ -1,51 +1,0 @@
-package com.sk.skala.axcalibur.spec.global.entity;
-
-import com.sk.skala.axcalibur.spec.global.entity.BaseTimeEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity
-@Table(name = "file_path")
-public class FilePathEntity extends BaseTimeEntity {
-    
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "`key`")
-    private Integer key;
-    
-    @Column(name = "path", nullable = false, unique = true, length = 100)
-    private String path;
-
-    @Column(name = "name", length = 255, nullable = false)
-    private String name;
-    
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "file_type_key", nullable = false)
-    private FileTypeEntity fileTypeKey;  // file_type 테이블 참조
-    
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "project_key", nullable = false)
-    private ProjectEntity projectKey;
-
-    // 같은 유형의 파일이 있으면 이름과 경로를 업데이트
-    public void updateFileInfo(String path, String name) {
-        this.path = path;
-        this.name = name;
-    }
-}
