--- backend/scenario/src/main/java/com/sk/skala/axcalibur/scenario/apilist/global/exception/BusinessExceptionHandler.java
@@ -1,26 +1,0 @@
-package com.sk.skala.axcalibur.spec.global.exception;
-
-import com.sk.skala.axcalibur.spec.global.code.ErrorCode;
-import lombok.Builder;
-import lombok.Getter;
-
-/**
- * 에러를 사용하기 위한 구현체
- */
-
-public class BusinessExceptionHandler extends RuntimeException{
-	@Getter
-	private final ErrorCode errorCode;
-	@Builder
-	public BusinessExceptionHandler(String message, ErrorCode errorCode) {
-		super(message);
-		this.errorCode = errorCode;
-	}
-
-	@Builder
-	public BusinessExceptionHandler(ErrorCode errorCode) {
-		super(errorCode.getMessage());
-		this.errorCode = errorCode;
-	}
-
-}
