--- backend/scenario/src/main/java/com/sk/skala/axcalibur/scenario/apilist/global/service/MultipartJackson2HttpMessageConverter.java
@@ -1,33 +1,0 @@
-package com.sk.skala.axcalibur.spec.global.service;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.lang.reflect.Type;
-import org.springframework.http.MediaType;
-import org.springframework.http.converter.json.AbstractJackson2HttpMessageConverter;
-import org.springframework.stereotype.Component;
-
-/**
- *  Swagger에서 파일 전송 시 에러 방지
- *
- * @author dig04214
- */
-@Component
-public class MultipartJackson2HttpMessageConverter extends AbstractJackson2HttpMessageConverter {
-	public  MultipartJackson2HttpMessageConverter(ObjectMapper objectMapper) {
-		super(objectMapper, MediaType.APPLICATION_OCTET_STREAM);
-	}
-	@Override
-	public boolean canWrite(Class<?> clazz, MediaType mediaType) {
-		return false;
-	}
-
-	@Override
-	public boolean canWrite(Type type, Class<?> clazz, MediaType mediaType) {
-		return false;
-	}
-
-	@Override
-	protected boolean canWrite(MediaType mediaType) {
-		return false;
-	}
-}
