+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/scenario/feature/apilist/dto/MappingRequestDto.java
@@ -1,0 +1,11 @@
+package com.sk.skala.axcalibur.scenario.apilist.feature.dto;
+
+import java.util.List;
+
+import lombok.Getter;
+
+@Getter
+public class MappingRequestDto {
+    private List<ScenarioDto> scenarioList;
+    private List<ApiDto> apiList;
+}
