+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/scenario/feature/apilist/dto/MappingResponseDto.java
@@ -1,0 +1,19 @@
+package com.sk.skala.axcalibur.scenario.apilist.feature.dto;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+import lombok.Getter;
+
+@Getter
+public class MappingResponseDto {
+    private String processedAt;
+    private Double validationRate;
+    private List<ApiMappingDto> apiMappingList;
+
+    public MappingResponseDto(String processedAt, Double validationRate, List<ApiMappingDto> apiMappingList) {
+        this.processedAt = processedAt;
+        this.validationRate = validationRate;
+        this.apiMappingList = apiMappingList;
+    }
+}
