+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/scenario/feature/apilist/entity/ScenarioEntity.java
@@ -1,0 +1,52 @@
+package com.sk.skala.axcalibur.scenario.feature.apilist.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+
+import com.sk.skala.axcalibur.scenario.global.entity.BaseTimeEntity;
+import com.sk.skala.axcalibur.scenario.global.entity.ProjectEntity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Entity
+@Table(name = "scenario")
+public class ScenarioEntity extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "`key`")
+    private Integer key;
+
+    @Column(name = "id", nullable = false, length = 30, unique = true)
+    private String id;
+
+    @Column(name = "name", nullable = false, length = 100)
+    private String name;
+
+    @Column(name = "description", nullable = false, columnDefinition = "TEXT")
+    private String description;
+
+    @Column(name = "validation", nullable = true)
+    private String validation;
+
+    @Column(name = "flow_chart", nullable = true)
+    private String flowChart;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "project_key", nullable = false)
+    private ProjectEntity projectKey;
+    
+    
+}
