+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/scenario/global/config/RedisConfig.java
@@ -1,0 +1,62 @@
+package com.sk.skala.axcalibur.scenario.global.config;
+// package com.sk.skala.axcalibur.sample.global.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
+import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
+import org.springframework.data.redis.serializer.RedisSerializer;
+import org.springframework.data.redis.serializer.StringRedisSerializer;
+
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+/**
+ * Resdis 연결 설정
+ *
+ * @author dig04214
+ */
+
+@RequiredArgsConstructor
+@EnableRedisRepositories(basePackages = "com.sk.skala.axcalibur.spec.feature.spec.repository")
+@Configuration
+@Slf4j
+public class RedisConfig {
+	@Value("${spring.data.redis.host}")
+	private String host;
+	@Value("${spring.data.redis.port}")
+	private int port;
+	@Value("${spring.data.redis.password}")
+	private String password;
+
+	@Bean
+	public RedisConnectionFactory redisConnectionFactory() {
+		RedisStandaloneConfiguration config= new RedisStandaloneConfiguration(host, port);
+		config.setPassword(password);
+
+		return new LettuceConnectionFactory(config);
+	}
+
+	@Bean
+	public RedisSerializer<Object> springSessionDefaultRedisSerializer() {
+		return new GenericJackson2JsonRedisSerializer();
+	}
+
+	@Bean
+	public RedisTemplate<String, Object> redisTemplate(
+		RedisConnectionFactory redisConnectionFactory,
+		RedisSerializer<Object> springSessionDefaultRedisSerializer
+	) {
+		RedisTemplate<String, Object> template = new RedisTemplate<>();
+		template.setConnectionFactory(redisConnectionFactory);
+		template.setKeySerializer(new StringRedisSerializer());
+		template.setValueSerializer(springSessionDefaultRedisSerializer);
+		return template;
+	}
+}
