+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/scenario/global/entity/BaseTimeEntity.java
@@ -1,0 +1,25 @@
+package com.sk.skala.axcalibur.scenario.global.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
+import lombok.Getter;
+import org.hibernate.annotations.Comment;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+/**
+ * 시간을 자동으로 채워주는 엔티티
+ *
+ * @author dig04214
+ */
+@Getter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public abstract class BaseTimeEntity {
+	@CreatedDate
+	@Column(updatable = false)
+	@Comment("생성일자")
+	private LocalDateTime createAt;
+}
